add seperate cdn file to serve
