+++ src/cdn-index.ts
@@ -1,0 +1,4 @@
+import initialize from '@core/index';
+
+// Attach to window/globalThis
+(globalThis as any).NCWebSDK = initialize;
